#18 - Applied PR reocmmendations. Fixed layout issues.
