+++ apps/middleman/src/app/(landing)/page.tsx
@@ -2,17 +2,15 @@ import Hero from "./components/Hero";
 import About from "./components/About";
 import EmptySpace from "./components/EmptySpace";
 import Services from "./components/Services";
-import Divider from "@/app/(landing)/components/Divider";
 
 export default function Landing() {
     return (
-        <>
-            <Divider top={510} />
+        <div className="h-[100vh] overflow-y-scroll scrollbar-hidden">
             <Hero />
             <About />
             <EmptySpace />
             <Services />
             <EmptySpace />
-        </>
+        </div>
     );
 }
