+++ packages/ui/src/styles/globals.css
@@ -259,3 +259,11 @@ input[type="number"]::-webkit-outer-spin-button {
   }
 
 }
+.scrollbar-hidden::-webkit-scrollbar {
+  display: none;
+}
+
+.scrollbar-hidden {
+  -ms-overflow-style: none;  /* IE and Edge */
+  scrollbar-width: none;  /* Firefox */
+}
